Remove .Core and just have the bare OpenLaw with core functionality
So we have a lib-only without CLI concerns.
