+++ src/OpenLaw.Commands/Argentina/DownloadCommand.cs
@@ -79,7 +79,7 @@ public class DownloadCommand(IAnsiConsole console, IHttpClientFactory http) : As
     }
 
     static void Convert(string file, bool overwrite)
-        => DictionaryConverter.Convert(file, true, true, true, overwrite);
+        => JsonDataConverter.Convert(file, true, true, true, overwrite);
 
     static async Task<long> GetJsonTimestampAsync(string file)
     {
