+++ src/OpenLaw.Commands/OpenLaw.Commands.csproj
@@ -23,7 +23,6 @@
     <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="8.0.1" />
     <PackageReference Include="Devlooped.Web" Version="1.3.0" />
     <PackageReference Include="NuGet.Protocol" Version="6.13.2" />
-    <PackageReference Include="NuGetizer" Version="1.2.4" PrivateAssets="all" />
     <PackageReference Include="Polly" Version="8.5.2" />
     <PackageReference Include="Spectre.Console.Cli" Version="0.50.0" />
     <PackageReference Include="Spectre.Console.Json" Version="0.50.0" />
@@ -37,7 +36,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\OpenLaw.Core\OpenLaw.Core.csproj" />
+    <ProjectReference Include="..\OpenLaw\OpenLaw.csproj" />
   </ItemGroup>
 
 </Project>
