+++ src/OpenLaw/DictionaryConverter.cs
@@ -3,22 +3,15 @@ using System.Text.Json;
 using System.Text.RegularExpressions;
 using YamlDotNet.Serialization;
 
-namespace Clarius.OpenLaw.Argentina;
+namespace Clarius.OpenLaw;
 
+public record Link(string Url, string Name);
+
+/// <summary>
+/// Provides conversion methods to and from the data format used by SAIJ.
+/// </summary>
 public static partial class DictionaryConverter
 {
-#if NET9_0_OR_GREATER
-    static readonly Lock sync = new();
-#else
-    static readonly object sync = new();
-#endif
-
-    static readonly JsonSerializerOptions options = new()
-    {
-        Converters = { new JsonDictionaryConverter() },
-    };
-
-
     static readonly ISerializer serializer = new SerializerBuilder()
         .WithTypeConverter(new YamlDictionaryConverter())
         .WithTypeConverter(new YamlListConverter())
@@ -26,81 +19,12 @@ public static partial class DictionaryConverter
         .ConfigureDefaultValuesHandling(DefaultValuesHandling.OmitNull | DefaultValuesHandling.OmitEmptyCollections)
         .Build();
 
-    static readonly IDeserializer deserializer = new DeserializerBuilder().Build();
-
-    public static void Convert(string jsonFile, bool yaml, bool pdf, bool markdown, bool overwrite)
+    static readonly JsonSerializerOptions options = new()
     {
-        var yamlDir = Path.Combine(Path.GetDirectoryName(jsonFile) ?? "", "yaml");
-        var yamlFile = Path.Combine(yamlDir, Path.ChangeExtension(Path.GetFileName(jsonFile), ".yaml"));
-        if (yaml)
-            Directory.CreateDirectory(yamlDir);
+        Converters = { new JsonDictionaryConverter() },
+    };
 
-        var mdDir = Path.Combine(Path.GetDirectoryName(jsonFile) ?? "", "md");
-        var mdFile = Path.Combine(mdDir, Path.ChangeExtension(Path.GetFileName(jsonFile), ".md"));
-        if (markdown)
-            Directory.CreateDirectory(mdDir);
-
-        var pdfDir = Path.Combine(Path.GetDirectoryName(jsonFile) ?? "", "pdf");
-        var pdfFile = Path.Combine(pdfDir, Path.ChangeExtension(Path.GetFileName(jsonFile), ".pdf"));
-        if (pdf)
-            Directory.CreateDirectory(pdfDir);
-
-        Dictionary<string, object?>? dictionary = null;
-        var writeTime = File.GetLastWriteTimeUtc(jsonFile);
-
-        if (overwrite || !File.Exists(yamlFile) ||
-            File.Exists(yamlFile) && File.GetLastWriteTimeUtc(yamlFile) != writeTime)
-        {
-            dictionary = Parse(File.ReadAllText(jsonFile));
-            if (dictionary is null)
-                return;
-
-            File.WriteAllText(yamlFile, dictionary.ToYaml(), Encoding.UTF8);
-        }
-
-        // Always ensure write time matches source json file
-        File.SetLastWriteTimeUtc(yamlFile, writeTime);
-
-        if (overwrite || !File.Exists(mdFile) ||
-            File.Exists(mdFile) && File.GetLastWriteTimeUtc(mdFile) != writeTime)
-        {
-            dictionary ??= Parse(File.ReadAllText(jsonFile));
-            if (dictionary is null)
-                return;
-
-            File.WriteAllText(mdFile, dictionary.ToMarkdown(), Encoding.UTF8);
-        }
-
-        // Always ensure write time matches source json file
-        File.SetLastWriteTimeUtc(mdFile, writeTime);
-
-        if (overwrite || !File.Exists(pdfFile) ||
-            File.Exists(pdfFile) && File.GetLastWriteTimeUtc(pdfFile) != writeTime)
-        {
-            dictionary ??= Parse(File.ReadAllText(jsonFile));
-            if (dictionary is null)
-                return;
-
-            if (!File.Exists(mdFile))
-            {
-                File.WriteAllText(mdFile, dictionary.ToMarkdown(), Encoding.UTF8);
-                File.SetLastWriteTimeUtc(mdFile, writeTime);
-            }
-
-            // single-threaded usage of the markdown2pdf converter
-            string? converted = null;
-            lock (sync)
-            {
-                // Force sync execution since we can't lock and await.
-                // Should work ok because it's a console app.
-                converted = new Markdown2Pdf.Markdown2PdfConverter().Convert(mdFile).Result;
-            }
-            File.Move(converted, pdfFile, overwrite: true);
-        }
-
-        // Always ensure write time matches source json file
-        File.SetLastWriteTimeUtc(pdfFile, writeTime);
-    }
+    static readonly IDeserializer deserializer = new DeserializerBuilder().Build();
 
     public static Dictionary<string, object?>? Parse(string json)
         => JsonSerializer.Deserialize<Dictionary<string, object?>>(json, options);
