+++ src/OpenLaw/JsonDictionaryConverter.cs
