+++ src/OpenLaw/OpenLaw.csproj
@@ -8,9 +8,8 @@
 
   <ItemGroup>
     <PackageReference Include="NuGetizer" Version="1.2.4" PrivateAssets="all" />
-    <PackageReference Include="Devlooped.TableStorage" Version="5.2.1" />
-    <PackageReference Include="Devlooped.WhatsApp" Version="1.0.0-rc.3" />
     <PackageReference Include="SystemTextJson.JsonDiffPatch" Version="2.0.0" />
+    <PackageReference Include="YamlDotNet" Version="16.3.0" />
   </ItemGroup>
 
 </Project>
