+++ src/Tests/Tests.csproj
@@ -18,10 +18,11 @@
     <PackageReference Include="Microsoft.Extensions.Configuration" Version="9.0.3" />
     <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="9.0.3" />
     <PackageReference Include="Microsoft.Extensions.Http" Version="9.0.3" />
+    <PackageReference Include="Devlooped.JQ" Version="1.7.1.8" />
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\OpenLaw.Core\OpenLaw.Core.csproj" />
+    <ProjectReference Include="..\OpenLaw\OpenLaw.csproj" />
     <ProjectReference Include="..\OpenLaw.Commands\OpenLaw.Commands.csproj" />
   </ItemGroup>
